+++ mod/src/RPPostUpdate.lua
@@ -70,6 +70,9 @@ function RPPostUpdate:Main()
     RPGlobals.run.queuedItems = true
     if player.QueuedItem.Item.Type == ItemType.ITEM_PASSIVE then -- 1
       RPGlobals.run.passiveItems[#RPGlobals.run.passiveItems + 1] = player.QueuedItem.Item.ID
+      if player.QueuedItem.Item.ID == CollectibleType.COLLECTIBLE_MUTANT_SPIDER_INNER_EYE then
+        Isaac.DebugString("Adding collectible 3001 (Mutant Spider's Inner Eye)")
+      end
       RPSpeedrun:CheckSeason5Start()
     end
   end